Source reorganization of Android `socket.h`
Create a bionic module and move some things over from `linux_like`.

Source: https://cs.android.com/android/platform/superproject/main/+/main:bionic/libc/include/sys/socket.h;drc=dc6e2fb4857f05bfee435cfd4f3859e74c746811

(backport <https://github.com/rust-lang/libc/pull/4541>)
(cherry picked from commit bce7db1b5d3ba2f10e1caa17e6401ebf8016cf32)
